Quick fix base
