+++ lib/src/endpoints/endpoints.dart
@@ -1,0 +1,9 @@
+Map<String, String> headers() {
+  Map<String, String> headers = {
+    'Content-Type': 'application/json',
+    'Accept': 'application/json',
+    'Authorization':
+        'Bearer eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJhZG1pbiIsImV4cCI6MTYyMjU5ODc5MywiaWF0IjoxNjIyNTk2OTkzfQ.jIj9-1Vz_h5xpg8NQRoPK0pHzihkLJ5bhpVQqiBgt5YL0LwoIsjGR86Lxdbqwbohi75Y0ntyf3Buq4Qlkf8gog',
+  };
+  return headers;
+}
