+++ lib/src/pages/notifications_page.dart
@@ -1,9 +1,14 @@
+import 'package:easystory/src/drawer.dart';
 import 'package:flutter/material.dart';
 
 class NotificationPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      appBar: AppBar(
+        title: Text('EasyStory'),
+      ),
+      drawer: MyDrawer(),
       body: Center(
         child: Text('No tienes notificaciones nuevas.'),
       ),
